requirements.txt
account==0.1.0
alembic==1.13.2
amqp==2.6.1
aniso8601==9.0.1
antiorm==1.2.1
anyio==3.7.1
api==0.0.7
app==0.0.1
apturl==0.5.2
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
blinker==1.8.2
books==0.1.2
Bootstrap-Flask==2.4.0
Brlapi==0.8.3
cachetools==5.5.0
celery==4.4.7
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cssselect==1.2.0
cssutils==2.11.1
cupshelpers==1.0
db==0.1.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.1
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-utils-six==2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
dominate==2.9.1
duplicity==0.8.21
ecdsa==0.19.0
email-validator==1.3.1
emails==0.5.15
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.70.1
fasteners==0.14.1
filelock==3.6.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
flask-shell-ipython==0.5.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
future==0.18.2
greenlet==3.0.3
gunicorn==20.1.0
h11==0.9.0
html5lib==1.1
httplib2==0.20.2
httptools==0.1.2
idna==3.3
importlib-metadata==4.6.4
inflect==5.6.2
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mock==5.1.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
nose==1.3.7
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==24.1
paramiko==2.9.3
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==2.5.1
pluggy==0.13.1
premailer==3.10.0
projects==1.0.9
prompt_toolkit==3.0.47
protobuf==3.12.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.0
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.18
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.9.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==5.4.3
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2024.2
pyxdg==0.27
PyYAML==5.4.1
raven==6.10.0
reportlab==3.6.8
requests==2.25.1
rsa==4.9
scour==0.38.2
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.34
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.16.0
Student==0.0.1
systemd-python==234
tenacity==6.3.1
termcolor==2.4.0
toml==0.10.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
urwid==2.6.15
usb-creator==0.3.7
uvicorn==0.11.8
uvloop==0.20.0
vine==1.3.0
virtualenv==20.13.0+ds
visitor==0.1.3
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websockets==8.1
Werkzeug==3.0.4
WTForms==3.1.2
xdg==5
xkit==0.0.0
zipp==1.0.0
